Make NsPerPixel a double instead of truncating (#2796)
As mentioned in https://github.com/google/orbit/pull/2795, we are using
an integer value for ns_per_pixel, which leads us to a problem because
if we multiply the number of pixels with ns_per_pixel we have a
different value than the total number of pixels.

In this PR, we are switching it to double and we cast to an int for
getting the correct timestamp.

Test: Load a capture, see than the image is the same than the current
main.

+++ src/OrbitGl/ThreadTrack.cpp
@@ -450,7 +450,7 @@ void ThreadTrack::OnCaptureComplete() {
   float normalized_x = (world_x - draw_data.track_start_x) / draw_data.track_width;
   int pixel_x = static_cast<int>(
       ceil(normalized_x * draw_data.viewport->WorldToScreenWidth(draw_data.track_width)));
-  return draw_data.min_tick + pixel_x * draw_data.ns_per_pixel;
+  return draw_data.min_tick + static_cast<uint64_t>(pixel_x * draw_data.ns_per_pixel);
 }
 
 // We minimize overdraw when drawing lines for small events by discarding events that would just
